[#5] - Started Implementation of transmission of shopID

+++ src/main/java/de/hpi/crawler/service/model/data/CrawledPage.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 
 @Getter @Setter
 public class CrawledPage {
-    int shopID;
+    long shopID;
     String url;
     String htmlSource;
 }
